Changed Resources type from int to float Changed display of resources to show 2 decimal places

+++ QuantumWorld_v1.0/Model/BuildingModel.cs
@@ -25,14 +25,13 @@ namespace QuantumWorld_v1._0.Model
             Level = level;
         }
 
-        public void SetNewCost(int index, int multiplier)
+        public void SetNewCost(int index, float multiplier)
         {
-           
+            
             Cost[index].MultiplyBy(multiplier);
-           
         }
-    
-        
+
+
         public void IncreaseLevel()
         {
             Level++;
